requirements.txt
streamlit>=1.36
streamlit-ace>=0.1.1
google-generativeai>=0.8
groq>=0.11
httpx>=0.27
pydantic>=2.7
pydantic-settings>=2.3
python-dotenv>=1.0
diskcache>=5.6
jinja2>=3.1
SQLAlchemy>=2.0
